Initial commit for 1 hour game jam web code
In this stage, updates to the site are manual, done by adding javascript function calls at the top of the file, as described in index.php. To run this version, only a web server is required, even the .php file can be renamed to a .html file, meaning even support for PHP is not needed.